♻️  Refactor Record Game into a component
